Biome & surface generation fixes (#648)
* add biome test

* tweak biome data

* add multi noise sample test

* multi noise work

* extract tree itself

* tweak tests

* chunk tests complete

* add surface test

* fix badlands

* refactor perlin noise

* frozen ocean tests

* fix glaciers

* cache optimization

* cache block lookup

* update asset files

* add second badlands test

* add biome blend test

* add more tests

+++ pumpkin/src/world/mod.rs
@@ -462,7 +462,7 @@ impl World {
                 false,
                 (self.dimension_type as u8).into(),
                 self.dimension_type.name(),
-                biome::hash_seed(self.level.seed.0 as i64), // seed
+                biome::hash_seed(self.level.seed.0), // seed
                 gamemode as u8,
                 base_config.default_gamemode as i8,
                 false,
@@ -765,7 +765,7 @@ impl World {
             .enqueue_packet(&CRespawn::new(
                 (self.dimension_type as u8).into(),
                 self.dimension_type.name(),
-                biome::hash_seed(self.level.seed.0 as i64), // seed
+                biome::hash_seed(self.level.seed.0), // seed
                 player.gamemode.load() as u8,
                 player.gamemode.load() as i8,
                 false,
